requirements.txt
fastapi>=0.104,<0.111
uvicorn[standard]>=0.24,<0.27
httpx>=0.25,<0.28
pyyaml>=6,<7
croniter>=1,<2
jinja2>=3.1,<3.2
